busybox: Disable sleep as ash builtin which breaks recovery
Busybox's ash has the option for sleep to a builtin. Builtin
in this context meaning as a command rather than as external program
that doesn't require a process to be created for each call to a
builtin command.

Sleep being run inside the shell process it self can cause various
issue, which is the reason upstream disabled the option.

In our case the bug builtin sleep triggered was caused by a later
commit which was supposed to fix a bug in builtin sleep but only
caused for the command to be interrupted by SIGCHILD.
For sleep to not well sleep caused in turn then for the shell to exit
before intended.

Because sleep didn't wait the recovery init exited which
caused the kernel to panic as init is not supposed to exit by itself.

Further sleep as builtin also causes ash to exit when the user passes
wrong arguments such as 1ab or 0.5.

[busybox] Disable sleep as ash builtin which breaks recovery. Fixes JB#61382

See-also: https://git.busybox.net/busybox/commit/?id=5e0411a7fb510b9aecda0a850c76bdd62c50efa4
See-also: https://bugs.busybox.net/show_bug.cgi?id=15619
Fixes: be6020a4aca08f1e30d73f7cf1ac20342adca1f4
Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>